Return artificial events to Handler for missing events
This fixes a bug, and prevents one in the new rerun-fails feature.

Incomplete tests were printed in the summary, but never passed to the EventHandler. Now when the execution ends, fake TestEVent will be created and sent to the Handler, so that it can print the output of the failed test.

This allows the rerun-fails loop to include these failed tests in the count, and avoid a situation where a test was panicing, but not getting rerun. If the panic test did not get rerun then 'go test' may exit 0, making it look like everything had passed, when in fact one test was still panicing.

This type of bug sounds pretty severe.